+++ resources/views/table/filter/filter-form.blade.php
@@ -1,3 +1,3 @@
-@use('Patrikjak\Utils\Table\Enums\Filter\FilterType')
+@use('Patrikjak\Utils\Common\Enums\Filter\FilterType')
 
 @include('pjutils::table.filter.filter-forms.' . $type->value)
