+++ src/Common/Dto/Filter/NumberFilterCriteria.php
@@ -2,9 +2,9 @@
 
 declare(strict_types = 1);
 
-namespace Patrikjak\Utils\Table\Dto\Filter\Criteria;
+namespace Patrikjak\Utils\Common\Dto\Filter;
 
-use Patrikjak\Utils\Table\Enums\Filter\FilterType;
+use Patrikjak\Utils\Common\Enums\Filter\FilterType;
 
 class NumberFilterCriteria extends BaseFilterCriteria
 {
