+++ src/Table/Dto/Filter/Definitions/FilterDefinition.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace Patrikjak\Utils\Table\Dto\Filter\Definitions;
 
-use Patrikjak\Utils\Table\Enums\Filter\FilterType;
+use Patrikjak\Utils\Common\Enums\Filter\FilterType;
 
 interface FilterDefinition
 {
