+++ src/Table/Dto/Filter/Definitions/Number/NumberFilterDefinition.php
@@ -5,7 +5,7 @@ declare(strict_types = 1);
 namespace Patrikjak\Utils\Table\Dto\Filter\Definitions\Number;
 
 use Patrikjak\Utils\Table\Dto\Filter\Definitions\FilterDefinition;
-use Patrikjak\Utils\Table\Enums\Filter\FilterType;
+use Patrikjak\Utils\Common\Enums\Filter\FilterType;
 
 class NumberFilterDefinition implements FilterDefinition
 {
