+++ src/Table/Dto/Filter/Definitions/Select/SelectFilterDefinition.php
@@ -6,7 +6,7 @@ namespace Patrikjak\Utils\Table\Dto\Filter\Definitions\Select;
 
 use Patrikjak\Utils\Table\Dto\Filter\Definitions\FilterDefinition;
 use Patrikjak\Utils\Table\Dto\Filter\Definitions\NeedsData;
-use Patrikjak\Utils\Table\Enums\Filter\FilterType;
+use Patrikjak\Utils\Common\Enums\Filter\FilterType;
 
 readonly class SelectFilterDefinition implements FilterDefinition, NeedsData
 {
