+++ src/Table/Dto/Filter/Definitions/Text/TextFilterDefinition.php
@@ -4,8 +4,8 @@ declare(strict_types = 1);
 
 namespace Patrikjak\Utils\Table\Dto\Filter\Definitions\Text;
 
+use Patrikjak\Utils\Common\Enums\Filter\FilterType;
 use Patrikjak\Utils\Table\Dto\Filter\Definitions\FilterDefinition;
-use Patrikjak\Utils\Table\Enums\Filter\FilterType;
 
 final readonly class TextFilterDefinition implements FilterDefinition
 {
