+++ src/Table/Dto/Filter/Settings.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace Patrikjak\Utils\Table\Dto\Filter;
 
-use Patrikjak\Utils\Table\Dto\Filter\Criteria\FilterCriteria;
+use Patrikjak\Utils\Common\Dto\Filter\FilterCriteria;
 use Patrikjak\Utils\Table\Dto\Filter\Definitions\FilterableColumn;
 
 final readonly class Settings
