+++ src/Table/Factories/Cells/CellFactory.php
@@ -2,11 +2,13 @@
 
 declare(strict_types = 1);
 
-namespace Patrikjak\Utils\Table\Dto\Cells;
+namespace Patrikjak\Utils\Table\Factories\Cells;
 
 use Patrikjak\Utils\Common\Enums\Icon;
 use Patrikjak\Utils\Common\Enums\Type;
+use Patrikjak\Utils\Table\Dto\Cells\Chip;
 use Patrikjak\Utils\Table\Dto\Cells\Double as DoubleCell;
+use Patrikjak\Utils\Table\Dto\Cells\Simple;
 
 readonly class CellFactory
 {
