+++ src/Table/Http/Controllers/TableController.php
@@ -6,7 +6,7 @@ namespace Patrikjak\Utils\Table\Http\Controllers;
 
 use Illuminate\Http\JsonResponse;
 use Illuminate\View\Compilers\BladeCompiler;
-use Patrikjak\Utils\Table\Enums\Filter\FilterType;
+use Patrikjak\Utils\Common\Enums\Filter\FilterType;
 use Patrikjak\Utils\Table\View\Filter\FilterForm;
 use Symfony\Component\HttpFoundation\Response;
 use Throwable;
