+++ src/Table/Http/Requests/TableParametersRequest.php
@@ -7,17 +7,17 @@ namespace Patrikjak\Utils\Table\Http\Requests;
 use Carbon\CarbonImmutable;
 use Illuminate\Cookie\CookieJar;
 use Illuminate\Foundation\Http\FormRequest;
-use Patrikjak\Utils\Table\Dto\Filter\Criteria\BaseFilterCriteria;
-use Patrikjak\Utils\Table\Dto\Filter\Criteria\DateFilterCriteria;
-use Patrikjak\Utils\Table\Dto\Filter\Criteria\FilterCriteria;
-use Patrikjak\Utils\Table\Dto\Filter\Criteria\NumberFilterCriteria;
-use Patrikjak\Utils\Table\Dto\Filter\Criteria\SelectFilterCriteria;
-use Patrikjak\Utils\Table\Dto\Filter\Criteria\TextFilterCriteria;
+use Patrikjak\Utils\Common\Dto\Filter\BaseFilterCriteria;
+use Patrikjak\Utils\Common\Dto\Filter\DateFilterCriteria;
+use Patrikjak\Utils\Common\Dto\Filter\FilterCriteria;
+use Patrikjak\Utils\Common\Dto\Filter\NumberFilterCriteria;
+use Patrikjak\Utils\Common\Dto\Filter\SelectFilterCriteria;
+use Patrikjak\Utils\Common\Dto\Filter\TextFilterCriteria;
+use Patrikjak\Utils\Common\Dto\Sort\SortCriteria;
+use Patrikjak\Utils\Common\Enums\Filter\FilterType;
+use Patrikjak\Utils\Common\Enums\Filter\TextFilterType;
+use Patrikjak\Utils\Common\Enums\Sort\SortOrder;
 use Patrikjak\Utils\Table\Dto\Parameters;
-use Patrikjak\Utils\Table\Dto\Sort\SortCriteria;
-use Patrikjak\Utils\Table\Enums\Filter\FilterType;
-use Patrikjak\Utils\Table\Enums\Filter\TextFilterType;
-use Patrikjak\Utils\Table\Enums\Sort\SortOrder;
 use stdClass;
 
 class TableParametersRequest extends FormRequest
