+++ src/Table/Services/Sortable.php
@@ -4,8 +4,8 @@ declare(strict_types = 1);
 
 namespace Patrikjak\Utils\Table\Services;
 
+use Patrikjak\Utils\Common\Dto\Sort\SortCriteria;
 use Patrikjak\Utils\Table\Dto\Sort\SortableColumn;
-use Patrikjak\Utils\Table\Dto\Sort\SortCriteria;
 
 interface Sortable extends Renderable
 {
