+++ src/Table/View/Sort/SortOption.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 
 namespace Patrikjak\Utils\Table\View\Sort;
 
-use Patrikjak\Utils\Table\Dto\Sort\SortCriteria;
+use Patrikjak\Utils\Common\Dto\Sort\SortCriteria;
 
 final readonly class SortOption
 {
