+++ tests/Unit/Common/Services/QueryBuilder/SortServiceTest.php
@@ -5,9 +5,9 @@ namespace Patrikjak\Utils\Tests\Unit\Common\Services\QueryBuilder;
 use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Database\DatabaseManager;
 use Orchestra\Testbench\TestCase;
+use Patrikjak\Utils\Common\Dto\Sort\SortCriteria;
+use Patrikjak\Utils\Common\Enums\Sort\SortOrder;
 use Patrikjak\Utils\Common\Services\QueryBuilder\SortService;
-use Patrikjak\Utils\Table\Dto\Sort\SortCriteria;
-use Patrikjak\Utils\Table\Enums\Sort\SortOrder;
 
 class SortServiceTest extends TestCase
 {
