+++ tests/Unit/Table/Factories/Cells/CellTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Patrikjak\Utils\Tests\Unit\Table\Dto\Cells;
+namespace Patrikjak\Utils\Tests\Unit\Table\Factories\Cells;
 
 use Patrikjak\Utils\Common\Enums\Icon;
 use Patrikjak\Utils\Common\Enums\Type;
 use Patrikjak\Utils\Table\Dto\Cells\Cell as AbstractCell;
-use Patrikjak\Utils\Table\Dto\Cells\CellFactory;
 use Patrikjak\Utils\Table\Enums\Cells\CellType;
+use Patrikjak\Utils\Table\Factories\Cells\CellFactory;
 use Patrikjak\Utils\Table\Interfaces\Cells\Cell as CellInterface;
 use PHPUnit\Framework\TestCase;
